Increase the test coverage of the Error classes.
The RequestError class accepts a Seq[Throwable], in case this is empty we should handle this accordingly and print a better message, therefore instead of a Seq we use a NonEmptyList.
The RequestError class is renamed to Multiple and a deprecated unapply method is put in the RequestError object to provide backwards compatibility.
The problem with invariant NonEmptyList is circumvented by using `Functor.widen` in the Endpoint.
Multi-platform lineSeparators are used.
A typo in the BenchmarkSpec is fixed.

Implement tests for missed implicit classes.
StringNelEndpointOps and StringSeqEndpointOps are both covered.

Update documentation regarding `Multiple`, and make the example compile.