Refactor the way how JSON Encoders for Exceptions are provided
The main goal of this commit is to delegate the process of
transforming Exceptions into JSON to the JSON libraries instead
of handling it in the core package. Thus JSON Encoder for
Exceptions is removed from the core package, but appropriate
instances are provided in each of subprojects responsible for
integration with the corresponding JSON libraries.

The motivation is that transforming even very simple data
structures into the valid JSON is not so trivial and at least
involves some logic of escaping strings before inserting them into
the JSON document. Also, it seems obvious, that in the situation
when exception need to be encoded into JSON, other data should be
JSON encoded as well, and user will end up using some JSON library
anyway. So, delegating this process (of encoding Exceptions) to
JSON libraries does not force users to use one of them, because
they already do. On the other hand it corresponds to the SoC
principle and make the code better.