Optimize Scala Future -> Twitter Future conversion
By using a custom CallingThreadExecutionContext we can save a trip
through the thread pool and two context switches.

This is an optimization that Twitter Future does by default, but for
Scala Future this needs to be done manually (where appropriate).

Also make ScalaToTwitterFuture a val instead of a def, for a decrease
in garbage creation and increase in performance.